TensorImpl: Move symbolic refresh_numel and refresh_contiguous into their own class
[ghstack-poisoned]
